fix: 全项目代码审查修复（27 项 bug 和优化）
P0（严重）：
- 修复 Encryption::verifySign() 致命错误（hash_equals 缺少参数），移除错放的 isWeekend
- CORS 中间件添加 Origin 白名单机制，修复任意域反射安全漏洞
- UserManageController 所有方法添加 #[Permission] 注解保护敏感操作
- UniApp 微信快捷登录/绑定手机后同步 userStore 状态
- PC article tags 类型从 string 修正为 string[]
- PC useRequest 中 navigateTo 替换为 window.location.href（修复脱离 Nuxt 上下文）

P1（重要）：
- BalanceLogRepository::existsBySource 添加 user_id 范围限定
- AdminService deleteAdmin/updateStatus 从参数传入 currentAdminId，移除 Service 层 request() 调用
- AdminPermissionMiddleware 在 PermissionSkip 分支也注入 userInfo
- Admin notification-bell setInterval 添加 onUnmounted 清理
- PC 支付成功后重置 recharging 状态；balance/points 分页参数修正为 page_no
- UniApp pay-result 将 checkPayResult 移入 onLoad 修复生命周期问题
- Admin workbench loadStats 添加 try/catch 防止白屏

P2（代码质量）：
- AdminRepository last_page 加 (int) 转换
- ConfigChangedListener 配置变更时调用 StorageManager::reset()
- Upload 组件批量上传完成检测优化
- multipleTabs removeTab 在仅剩一个标签时提前返回
- PC article 页面移除重复 API 请求，复用已有数据
- PC useRequest handleResponse 改为单例 message 实例
- PC user/index 改用 definePageMeta redirect
- PC store/user 添加 import.meta.client 守卫
- UniApp 版本号改为动态读取
- UniApp formatDate 统一使用 utils/time
- UniApp getClientType 添加 fallback 返回值
- UniApp d-phone-login 和 useLogin 统一迁移到 useCountdown
- UniApp usePagingList refresh 添加 catch 处理

+++ server/app/repository/system/AdminRepository.php
@@ -62,7 +62,7 @@ class AdminRepository extends Repository
                 'current_page' => $page,
                 'per_page' => $limit,
                 'total' => $total,
-                'last_page' => ceil($total / $limit)
+                'last_page' => (int) ceil($total / $limit)
             ]
         ];
     }
